[android] fix: Perspective view settings moved to Routing group. del: Removed "Will drain battery" subtitle.
