+++ android/res/xml-v21/prefs_map.xml
@@ -32,23 +32,15 @@
     android:order="4"/>
 
   <SwitchPreference
-    android:key="@string/pref_3d"
-    android:title="@string/pref_map_3d_title"
+    android:key="@string/pref_3d_buildings"
+    android:title="@string/pref_map_3d_buildings_title"
     android:switchTextOn=""
     android:switchTextOff=""
     android:order="5"/>
-
-  <SwitchPreference
-    android:key="@string/pref_3d_buildings"
-    android:title="@string/pref_map_3d_buildings_title"
-    android:summary="@string/pref_map_3d_buildings_subtitle"
-    android:switchTextOn=""
-    android:switchTextOff=""
-    android:order="6"/>
 
   <Preference
     android:key="@string/pref_yota"
     android:title="@string/yopme_pref_title"
     android:summary="@string/yopme_pref_summary"
-    android:order="7"/>
+    android:order="6"/>
 </PreferenceScreen>
