+++ android/res/xml/prefs_map.xml
@@ -30,19 +30,13 @@
     android:order="4"/>
 
   <CheckBoxPreference
-    android:key="@string/pref_3d"
-    android:title="@string/pref_map_3d_title"
-    android:order="5"/>
-
-  <CheckBoxPreference
     android:key="@string/pref_3d_buildings"
     android:title="@string/pref_map_3d_buildings_title"
-    android:summary="@string/pref_map_3d_buildings_subtitle"
-    android:order="6"/>
+    android:order="5"/>
 
   <Preference
     android:key="@string/pref_yota"
     android:title="@string/yopme_pref_title"
     android:summary="@string/yopme_pref_summary"
-    android:order="7"/>
+    android:order="6"/>
 </PreferenceScreen>
