+++ android/res/xml/prefs_route.xml
@@ -9,4 +9,8 @@
   <ListPreference
     android:key="@string/pref_tts_language"
     android:title="@string/pref_tts_language_title"/>
+
+  <CheckBoxPreference
+    android:key="@string/pref_3d"
+    android:title="@string/pref_map_3d_title"/>
 </PreferenceScreen>
