+++ android/src/com/mapswithme/maps/NewsFragment.java
@@ -50,7 +50,7 @@ public class NewsFragment extends BaseMwmDialogFragment
       images.recycle();
 
       // TODO: Temporary solution. Remove for the next WhatsNews
-      mSubtitles[0] += "\n\n" + MwmApplication.get().getString(R.string.whats_new_3d_update_maps);
+      mSubtitles[1] += "\n\n" + MwmApplication.get().getString(R.string.whats_new_3d_update_maps);
     }
 
     @Override
